+++ src/main/java/com/sparta/team_1_hyogeunchild/business/dto/OrderResponseDto.java
@@ -11,10 +11,6 @@ public class OrderResponseDto {
 
     private Product product;
 
-//    @ManyToOne
-//    @JoinColumn(name = "User_Id")
-//    private User user;
-
     private OrderResponseDto(Order order) {
         this.totalPrice = order.getTotalPrice();
         this.amount = order.getAmount();
